Add files via upload
This is a mini project in C. For this project I took Player as an entity and player details included player id(jersey number), player name, number of matches played, runs scored and wickets took by player. Player id is to be kept unique and hence a function is also written to validate it. This project even uses various String build-in functions.